- Sucessfully exported database - Populated exported database file - Imported the file - Added back the .createFromAsset in MainActivity.onCreate db = (...) - Changed name of database instance from "shoparound.db" to "usershoparound.db" in MainActivity.onCreate db = (...) to avoid any confusion with the name of the asset. - Please be aware of the item names in the database when testing. "apples" is an item in the database, "Apple" is not. In continued development after this course, accounting for this sort of thing should be a priority. I think we could alias "apple" to "apples", or possibly just include both as RetailItems linked to the same itemID.
Next up for me (EB) is exploring asynchronous queries. Since the workaround is working, we may do without for the submitted product, but I do want to see if it will be possible, since it is the "right way". In either case we should make note of it in the "technical considerations" section of our presentation.